テーブル設計変更　model seeder等作り直し。

+++ template-3rd-study/database/migrations/2021_12_12_003420_create_choices_table.php
@@ -15,7 +15,7 @@ class CreateChoicesTable extends Migration
     {
         Schema::create('choices', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('question_id');
+            $table->unsignedBigInteger('question_id');
             $table->string('name');
             $table->integer('valid');
             $table->timestamps();
